+++ Changelog.md
@@ -25,6 +25,7 @@
 - Change: Improved visuals when holding a taser
 - Change: Some alarm options have been moved to a separate screen
 - Change: The minimum radius of the Portable Radar is now one block
+- Change: The maximum allowed signature for a Keycard Reader is now 99999
 - API: IModuleInventory#getModuleDescriptionId to make it possible to have shared descriptions
 - API: Renamed Option#readFromNBT to Option#load and Option#writeToNBT to Option#save
 - Fix: Jade does not properly hide blocks